requirements.txt
beautifulsoup4==4.11.1
numpy==1.24.0
Pillow==9.3.0
PyYAML==6.0
requests==2.28.1
setuptools==65.5.0
